requirements.txt
dnspython>=1.16.0,<2.0.0
cryptography>=2.9.2
future
requests[security]
certbot-dns-hetzner